+++ app/src/main/java/com/yapp/gallery/MainActivity.kt
@@ -15,15 +15,9 @@ import javax.inject.Inject
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     @Inject lateinit var auth: FirebaseAuth
-    @Inject lateinit var sharedPreferences: SharedPreferences
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        // 토큰 매 번 받기
-        auth.currentUser?.getIdToken(false)?.addOnCompleteListener {
-            sharedPreferences.edit().putString("idToken", it.result.token).apply()
-        }
 
         setContentView(R.layout.activity_main)
         startLogin()
