+++ feature/profile/src/main/java/com/yapp/gallery/profile/screen/profile/ProfileActivity.kt
@@ -1,9 +1,11 @@
-package com.yapp.gallery.profile.screen
+package com.yapp.gallery.profile.screen.profile
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.yapp.gallery.common.theme.GalleryTheme
+import com.yapp.gallery.profile.screen.category.CategoryManageActivity
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -12,8 +14,12 @@ class ProfileActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             GalleryTheme {
-                ProfileScreen(popBackStack = { finish() })
+                ProfileScreen(popBackStack = { finish() }, navigateToManage = { navigateToManage() })
             }
         }
+    }
+
+    private fun navigateToManage(){
+        startActivity(Intent(this, CategoryManageActivity::class.java))
     }
 }
